+++ CHANGES.md
@@ -2,9 +2,9 @@
 
 ### Version 1.4.2:
 - Fixed a fault that allowed HTML to be injected into the display for events, tooltips, overlays, etc.
-- Added "Visible Days..." drop down menu support to the Day headers in the main display (opens the "Configuration" dialog).
+- Added "Visible Days..." drop-down menu support to the Day headers in the main display (opens the "Configuration" dialog).
 - Added new setting "useTemplateWhenAddingNewEvent" which adds a template event when "Add Event" is pressed (defaults to true).
-- 
+- Sorry for the long pause between releases...
 
 ### Version 1.4.1:
 - The Search dialogs text field is no longer cleared when closed and re-opened.  It is now focused and all the text is selected.